backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
duckdb==1.0.0
pydantic==2.9.2
python-dotenv==1.0.1
sqlmodel==0.0.21
pyarrow==17.0.0
pandas==2.2.2
matplotlib==3.9.2
requests==2.32.3
openai>=1.30.0
email-validator>=2.0.0
bcrypt
PyJWT
